refactor(frontend): extract useAvatar and useSettings hooks
- Decouple complex Audio context, WebSocket, and side-effect logic from Avatar.tsx into reusable useAvatar hook.
- Move useSettingsState custom hook to lib/hooks directory for better separation of concerns.
- Clean up massive component files, conforming to React best practices for Container/Presenter pattern.
